+++ fj-core/src/main/java/org/fugerit/java/core/io/helper/HelperIOException.java
@@ -79,6 +79,14 @@ public class HelperIOException extends IOException {
 		apply(fun, EX_HANDLER_SILENT);
 	}
 		
+	public static <T, E extends Exception> T getWithMessage( UnsafeSupplier<T, E> fun, String message ) throws IOException {
+		return get( fun, e -> { throw convertEx( message , e); } );
+	}
+	
+	public static <E extends Exception> void applyWithMessage( UnsafeVoid<E> fun, String message ) throws IOException {
+		apply(fun,  e -> { throw convertEx( message , e); } );
+	}
+	
 	public static <T, E extends Exception> T get( UnsafeSupplier<T, E> fun, UnsafeConsumer<Exception, IOException> exHandler ) throws IOException {
 		T res = null;
 		try {