+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- applyWithMessage() and getWithMessage() special function to ConfigException, DAOException and HelperIOException
+
 ## [8.4.2] - 2023-10-22
 
 ### Added
@@ -19,8 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - LICENSE place holder
-
-## [Unreleased]
 
 ## [8.4.1] - 2023-10-01